Plug Maven validator into the loop
